requirements.txt
langgraph
langchain
langchain-google-genai
langchain-mcp-adapters
python-dotenv
httpx


